feat: add language, framework, and runtime enums for enhanced categorization
- Introduced `LanguageKind`, `FrameworkKind`, `FileRole`, `ProgrammingParadigm`, and `RuntimeKind` enums to categorize programming languages, frameworks, file roles, paradigms, and runtime environments.
- Implemented `as_id` methods for each enum to provide string representations.

feat: implement console output for baseline reports

- Created `render_baseline_block` function to format baseline scan reports for console output.

feat: add findings rendering for console output

- Developed `render_grouped_findings` and `render_finding` functions to display findings in a structured format in the console.

feat: add React Native section rendering for console output

- Implemented `render_react_native_section` to display React Native project details in console output.

feat: create sections for console output

- Added `render_header`, `render_risk_summary`, `render_top_rules`, `render_top_risk_clusters`, `render_languages_section`, `render_frameworks_section`, and `render_framework_projects_section` functions for comprehensive console reporting.

feat: implement workspace risk summary for console output

- Developed `workspace_risk_table` function to summarize risks across workspace packages in console output.

feat: add markdown rendering for baseline reports

- Created `render_baseline_section` for formatting baseline reports in markdown.

feat: implement findings rendering for markdown output

- Developed `render_findings_index` and `render_grouped_findings` functions to display findings in markdown format.

feat: add React Native section rendering for markdown output

- Implemented `render_react_native_section` to display React Native project details in markdown.

feat: create sections for markdown output

- Added `render_overview`, `render_risk_summary`, `render_top_rules`, `render_top_risk_clusters`, `render_languages_section`, `render_frameworks_section`, and `render_framework_projects_section` functions for comprehensive markdown reporting.

feat: implement workspace risk summary for markdown output

- Developed `render_workspace_risk_table` function to summarize risks across workspace packages in markdown.
